# swagger-api.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cocktail Ordering System API",
    "version": "1.0.0",
    "description": "A complete backend API for a cocktail ordering system with Express.js, Node.js, and MongoDB. Features JWT authentication, Paystack payment integration, and comprehensive admin management.",
    "contact": {
      "name": "Clyp AI",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    },
    "schemas": {
...
      "OrderTracking": {
        "type": "object",
        "properties": {
          "orderNumber": {
            "type": "string",
            "description": "Order number"
          },
          "status": {
            "type": "string",
            "enum": [
              "new",
              "preparing",
              "delivered",
              "cancelled"
            ],
            "description": "Current order status"
          },
          "paymentStatus": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "failed",
              "refunded"
            ],
            "description": "Payment status"
          },
          "estimatedDelivery": {
            "type": "string",
            "format": "date-time",
            "description": "Estimated delivery time"
          },
          "orderProgress": {
            "type": "object",
            "properties": {
              "placed": {
                "type": "boolean",
                "description": "Order has been placed"
              },
              "confirmed": {
                "type": "boolean",
                "description": "Order has been confirmed"
              },
              "preparing": {
                "type": "boolean",
                "description": "Order is being prepared"
              },
              "delivered": {
                "type": "boolean",
                "description": "Order has been delivered"
              }
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "string"
                },
                "timestamp": {
                  "type": "string",
                  "format": "date-time"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/tracking/order/{orderNumber}": {
      "get": {
        "summary": "Track order status and progress (Public endpoint)",
        "tags": [
          "Order Tracking"
        ],
        "description": "Get detailed tracking information for an order including status, progress, and estimated delivery time.\nThis endpoint is perfect for customers to check their order status.\n",
        "parameters": [
          {
            "in": "path",
            "name": "orderNumber",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Order number (e.g., ORD-000001)",
            "example": "ORD-000001"
          }
        ],
        "responses": {
          "200": {
            "description": "Order tracking information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "orderNumber": {
                      "type": "string"
                    },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "state": {
                          "type": "string"
                        }
                      }
                    },
                    "tracking": {
                      "$ref": "#/components/schemas/OrderTracking"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "cocktail": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string"
                              },
                              "description": {
                                "type": "string"
                              },
                              "image": {
                                "type": "object"
                              }
                            }
                          },
                          "quantity": {
                            "type": "integer"
                          },
                          "price": {
                            "type": "number"
                          }
                        }
                      }
                    },
                    "totalAmount": {
                      "type": "number"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/tracking/customer/{phone}": {
      "get": {
        "summary": "Get all orders for a customer by phone number (Public endpoint)",
        "tags": [
          "Order Tracking"
        ],
        "description": "Retrieve all orders placed by a customer using their phone number.\nThis is useful for customers who want to see their order history.\n",
        "parameters": [
          {
            "in": "path",
            "name": "phone",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer phone number",
            "example": "+2348012345678"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "description": "Number of orders to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer orders retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "customer": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "totalOrders": {
                          "type": "integer"
                        },
                        "totalSpent": {
                          "type": "number"
                        }
                      }
                    },
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "orderNumber": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "paymentStatus": {
                            "type": "string"
                          },
                          "totalAmount": {
                            "type": "number"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "items": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No orders found for this phone number"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}